+++ accumulo/src/main/scala/geotrellis/spark/io/accumulo/AccumuloLayerProvider.scala
@@ -18,10 +18,10 @@ package geotrellis.spark.io.accumulo
 
 import geotrellis.spark._
 import geotrellis.spark.io._
+import geotrellis.spark.io.accumulo.conf.AccumuloConfig
 import geotrellis.util.UriUtils
 import org.apache.spark.SparkContext
-import org.apache.accumulo.core.client.security.tokens.PasswordToken
-import com.typesafe.config.ConfigFactory
+
 import java.net.URI
 
 /**
@@ -42,8 +42,7 @@ class AccumuloLayerProvider extends AttributeStoreProvider
   def attributeStore(uri: URI): AttributeStore = {
     val instance = AccumuloInstance(uri)
     val params = UriUtils.getParams(uri)
-    val attributeTable = params.getOrElse("attributes",
-      ConfigFactory.load().getString("geotrellis.accumulo.catalog"))
+    val attributeTable = params.getOrElse("attributes", AccumuloConfig.catalog)
     AccumuloAttributeStore(instance, attributeTable)
   }