+++ cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraInstance.scala
@@ -16,11 +16,11 @@
 
 package geotrellis.spark.io.cassandra
 
+import geotrellis.spark.io.cassandra.conf.CassandraConfig
+
 import com.datastax.driver.core.policies.{DCAwareRoundRobinPolicy, TokenAwarePolicy}
 import com.datastax.driver.core.{Cluster, Session}
-import com.typesafe.config.ConfigFactory
 
-import scala.util.Try
 import java.net.URI
 
 object CassandraInstance {
@@ -41,8 +41,7 @@ object CassandraInstance {
 }
 
 trait CassandraInstance extends Serializable {
-  val replicationStrategy: String
-  val replicationFactor: Int
+  val cassandraConfig: CassandraConfig
 
   /** Functions to get cluster / session for custom logic, where function wrapping can have an impact on speed */
   def getCluster: () => Cluster
@@ -62,7 +61,10 @@ trait CassandraInstance extends Serializable {
   }
 
   def ensureKeyspaceExists(keyspace: String, session: Session): Unit =
-    session.execute(s"create keyspace if not exists ${keyspace} with replication = {'class': '${replicationStrategy}', 'replication_factor': ${replicationFactor} }")
+    session.execute(
+      s"create keyspace if not exists ${keyspace} with replication = {'class': '${cassandraConfig.replicationStrategy}', " +
+      s"'replication_factor': ${cassandraConfig.replicationFactor} }"
+    )
 
   def dropKeyspace(keyspace: String, session: Session): Unit =
     session.execute(s"drop keyspace if exists $keyspace;")
@@ -87,64 +89,97 @@ trait CassandraInstance extends Serializable {
 
 case class BaseCassandraInstance(
   getCluster: () => Cluster,
-  replicationStrategy: String = Cassandra.cfg.getString("replicationStrategy"),
-  replicationFactor: Int = Cassandra.cfg.getInt("replicationFactor")
+  cassandraConfig: CassandraConfig
 ) extends CassandraInstance
 
 object BaseCassandraInstance {
+  def apply(getCluster: () => Cluster): BaseCassandraInstance =
+    BaseCassandraInstance(getCluster, CassandraConfig)
+
   def apply(
     hosts: Seq[String],
-    username: String = "",
-    password: String = "",
-    port: Int = Try(Cassandra.cfg.getInt("port")).toOption.getOrElse(9042),
-    replicationStrategy: String = Cassandra.cfg.getString("replicationStrategy"),
-    replicationFactor: Int = Cassandra.cfg.getInt("replicationFactor"),
-    localDc: String = Cassandra.cfg.getString("localDc"),
-    usedHostsPerRemoteDc: Int = Cassandra.cfg.getInt("usedHostsPerRemoteDc"),
-    allowRemoteDCsForLocalConsistencyLevel: Boolean = Cassandra.cfg.getBoolean("allowRemoteDCsForLocalConsistencyLevel")): BaseCassandraInstance = {
-
+    username: String,
+    password: String,
+    cassandraConfig: CassandraConfig
+  ): BaseCassandraInstance = {
     def getLoadBalancingPolicy = {
       val builder = DCAwareRoundRobinPolicy.builder()
-      if(localDc.nonEmpty) builder.withLocalDc(localDc)
-      if(usedHostsPerRemoteDc > 0) builder.withUsedHostsPerRemoteDc(0)
-      if(allowRemoteDCsForLocalConsistencyLevel) builder.allowRemoteDCsForLocalConsistencyLevel()
+      if(cassandraConfig.localDc.nonEmpty) builder.withLocalDc(cassandraConfig.localDc)
+      if(cassandraConfig.usedHostsPerRemoteDc > 0) builder.withUsedHostsPerRemoteDc(cassandraConfig.usedHostsPerRemoteDc)
+      if(cassandraConfig.allowRemoteDCsForLocalConsistencyLevel) builder.allowRemoteDCsForLocalConsistencyLevel()
 
       new TokenAwarePolicy(builder.build())
     }
 
-    val getCluster = () => Cluster.builder().withLoadBalancingPolicy(getLoadBalancingPolicy).addContactPoints(hosts: _*).withPort(port).build()
+    def getCluster = () => {
+      val builder =
+        Cluster
+          .builder()
+          .withLoadBalancingPolicy(getLoadBalancingPolicy)
+          .addContactPoints(hosts: _*)
+          .withPort(cassandraConfig.port)
+          .withCredentials(username, password)
 
-    BaseCassandraInstance(getCluster, replicationStrategy, replicationFactor)
+      val authedBuilder =
+        if(username.nonEmpty && password.nonEmpty) builder.withCredentials(username, password)
+        else builder
+
+      authedBuilder.build()
+    }
+
+
+    BaseCassandraInstance(getCluster, cassandraConfig)
   }
+
+  def apply(
+    hosts: Seq[String],
+    username: String,
+    password: String
+  ): BaseCassandraInstance = apply(hosts, username, password, CassandraConfig)
+
+  def apply(
+    hosts: Seq[String],
+    cassandraConfig: CassandraConfig
+  ): BaseCassandraInstance = apply(hosts, "", "", cassandraConfig)
+
+  def apply(hosts: Seq[String]): BaseCassandraInstance = apply(hosts, CassandraConfig)
 }
 
 object Cassandra {
-  lazy val cfg = ConfigFactory.load().getConfig("geotrellis.cassandra")
-
   implicit def instanceToSession[T <: CassandraInstance](instance: T): Session = instance.session
 
   def withCassandraInstance[T <: CassandraInstance, K](instance: T)(block: T => K): K = block(instance)
   def withCassandraInstanceDo[T <: CassandraInstance, K](instance: T)(block: T => K): K = try block(instance) finally instance.closeAsync
+
   def withBaseCassandraInstance[K](hosts: Seq[String],
-                                   username: String = "",
-                                   password: String = "",
-                                   port: Int = Try(Cassandra.cfg.getInt("port")).toOption.getOrElse(9042),
-                                   replicationStrategy: String = Cassandra.cfg.getString("replicationStrategy"),
-                                   replicationFactor: Int = Cassandra.cfg.getInt("replicationFactor"),
-                                   localDc: String = Cassandra.cfg.getString("localDc"),
-                                   usedHostsPerRemoteDc: Int = Cassandra.cfg.getInt("usedHostsPerRemoteDc"),
-                                   allowRemoteDCsForLocalConsistencyLevel: Boolean = Cassandra.cfg.getBoolean("allowRemoteDCsForLocalConsistencyLevel"))(block: CassandraInstance => K): K =
-    block(BaseCassandraInstance(hosts, username, password, port, replicationStrategy, replicationFactor, localDc, usedHostsPerRemoteDc, allowRemoteDCsForLocalConsistencyLevel))
+                                   username: String,
+                                   password: String,
+                                   cassandraConfig: CassandraConfig)(block: CassandraInstance => K): K =
+    block(BaseCassandraInstance(hosts, username, password, cassandraConfig))
+  def withBaseCassandraInstance[K](hosts: Seq[String], cassandraConfig: CassandraConfig)(block: CassandraInstance => K): K =
+    withBaseCassandraInstance(hosts, "", "", cassandraConfig)(block)
+  def withBaseCassandraInstance[K](hosts: Seq[String],
+                                   username: String,
+                                   password: String)(block: CassandraInstance => K): K =
+    withBaseCassandraInstance(hosts, username, password, CassandraConfig)(block)
+  def withBaseCassandraInstance[K](hosts: Seq[String])(block: CassandraInstance => K): K =
+    withBaseCassandraInstance(hosts, "", "", CassandraConfig)(block)
+
   def withBaseCassandraInstanceDo[K](hosts: Seq[String],
-                                     username: String = "",
-                                     password: String = "",
-                                     port: Int = Try(Cassandra.cfg.getInt("port")).toOption.getOrElse(9042),
-                                     replicationStrategy: String = Cassandra.cfg.getString("replicationStrategy"),
-                                     replicationFactor: Int = Cassandra.cfg.getInt("replicationFactor"),
-                                     localDc: String = Cassandra.cfg.getString("localDc"),
-                                     usedHostsPerRemoteDc: Int = Cassandra.cfg.getInt("usedHostsPerRemoteDc"),
-                                     allowRemoteDCsForLocalConsistencyLevel: Boolean = Cassandra.cfg.getBoolean("allowRemoteDCsForLocalConsistencyLevel"))(block: CassandraInstance => K): K = {
-    val instance = BaseCassandraInstance(hosts, username, password, port, replicationStrategy, replicationFactor, localDc, usedHostsPerRemoteDc, allowRemoteDCsForLocalConsistencyLevel)
+                                     username: String,
+                                     password: String,
+                                     cassandraConfig: CassandraConfig)(block: CassandraInstance => K): K = {
+    val instance = BaseCassandraInstance(hosts, username, password, cassandraConfig)
     try block(instance) finally instance.closeAsync
+  }
+  def withBaseCassandraInstanceDo[K](hosts: Seq[String],
+                                     username: String,
+                                     password: String)(block: CassandraInstance => K): K =
+    withBaseCassandraInstanceDo(hosts, username, password, CassandraConfig)(block)
+  def withBaseCassandraInstanceDo[K](hosts: Seq[String])(block: CassandraInstance => K): K =
+    withBaseCassandraInstanceDo(hosts, "", "", CassandraConfig)(block)
+  def withBaseCassandraInstanceDo[K](hosts: Seq[String],
+                                     cassandraConfig: CassandraConfig)(block: CassandraInstance => K): K = {
+    withBaseCassandraInstanceDo(hosts, "", "", cassandraConfig)(block)
   }
 }