+++ hbase/src/main/scala/geotrellis/spark/io/hbase/HBaseLayerProvider.scala
@@ -18,8 +18,9 @@ package geotrellis.spark.io.hbase
 
 import geotrellis.spark._
 import geotrellis.spark.io._
+import geotrellis.spark.io.hbase.conf.HBaseConfig
 import geotrellis.util.UriUtils
-import com.typesafe.config.ConfigFactory
+
 import org.apache.spark.SparkContext
 import java.net.URI
 
@@ -41,8 +42,7 @@ class HBaseLayerProvider extends AttributeStoreProvider
   def attributeStore(uri: URI): AttributeStore = {
     val instance = HBaseInstance(uri)
     val params = UriUtils.getParams(uri)
-    val attributeTable = params.getOrElse("attributes",
-      ConfigFactory.load().getString("geotrellis.hbase.catalog"))
+    val attributeTable = params.getOrElse("attributes", HBaseConfig.catalog)
     HBaseAttributeStore(instance, attributeTable)
   }