+++ s3/src/main/scala/geotrellis/spark/io/s3/S3RDDReader.scala
@@ -18,12 +18,12 @@ package geotrellis.spark.io.s3
 
 import geotrellis.spark._
 import geotrellis.spark.io._
+import geotrellis.spark.io.s3.conf.S3Config
 import geotrellis.spark.io.avro.codecs.KeyValueRecordCodec
 import geotrellis.spark.io.index.{IndexRanges, MergeQueue}
 import geotrellis.spark.io.avro.{AvroEncoder, AvroRecordCodec}
 import geotrellis.spark.util.KryoWrapper
 
-import com.typesafe.config.ConfigFactory
 import com.amazonaws.services.s3.model.AmazonS3Exception
 import org.apache.avro.Schema
 import org.apache.commons.io.IOUtils
@@ -31,8 +31,7 @@ import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 
 trait S3RDDReader {
-  final val DefaultThreadCount =
-    ConfigFactory.load().getThreads("geotrellis.s3.threads.rdd.read")
+  final val defaultThreadCount = S3Config.threads.rdd.readThreads
 
   def getS3Client: () => S3Client
 
@@ -47,7 +46,7 @@ trait S3RDDReader {
     filterIndexOnly: Boolean,
     writerSchema: Option[Schema] = None,
     numPartitions: Option[Int] = None,
-    threads: Int = DefaultThreadCount
+    threads: Int = defaultThreadCount
   )(implicit sc: SparkContext): RDD[(K, V)] = {
     if (queryKeyBounds.isEmpty) return sc.emptyRDD[(K, V)]