+++ s3/src/main/scala/geotrellis/spark/io/s3/cog/S3COGLayerWriter.scala
@@ -24,7 +24,7 @@ class S3COGLayerWriter(
   bucket: String,
   keyPrefix: String,
   getS3Client: () => S3Client = () => S3Client.DEFAULT,
-  threads: Int = S3RDDWriter.DefaultThreadCount
+  threads: Int = S3RDDWriter.defaultThreadCount
 ) extends COGLayerWriter {
 
   def writeCOGLayer[