+++ s3/src/main/scala/geotrellis/spark/io/s3/geotiff/S3GeoTiffLayerReader.scala
@@ -1,15 +1,13 @@
 package geotrellis.spark.io.s3.geotiff
 
 import geotrellis.spark.io.s3.cog.byteReader
+import geotrellis.spark.io.s3.conf.S3Config
 import geotrellis.raster.resample.{NearestNeighbor, ResampleMethod}
 import geotrellis.raster.io.geotiff.{AutoHigherResolution, OverviewStrategy}
 import geotrellis.spark.tiling.ZoomedLayoutScheme
 import geotrellis.spark.io.hadoop.geotiff.{AttributeStore, GeoTiffLayerReader, GeoTiffMetadata}
 import geotrellis.util.ByteReader
-import geotrellis.spark.io.ThreadConfig
 import geotrellis.spark.io.s3.S3Client
-
-import com.typesafe.config.ConfigFactory
 
 import java.net.URI
 
@@ -25,5 +23,5 @@ case class S3GeoTiffLayerReader[M[T] <: Traversable[T]](
 }
 
 object S3GeoTiffLayerReader {
-  val defaultThreadCount: Int = ConfigFactory.load().getThreads("geotrellis.s3.threads.collection.read")
+  val defaultThreadCount: Int = S3Config.threads.collection.readThreads
 }