+++ spark/src/main/scala/geotrellis/spark/io/file/geotiff/FileGeoTiffLayerReader.scala
@@ -3,13 +3,12 @@ package geotrellis.spark.io.file.geotiff
 import geotrellis.raster.resample.{NearestNeighbor, ResampleMethod}
 import geotrellis.raster.io.geotiff.{AutoHigherResolution, OverviewStrategy}
 import geotrellis.spark.tiling.ZoomedLayoutScheme
-import geotrellis.spark.io.ThreadConfig
 import geotrellis.spark.io.hadoop.geotiff._
 import geotrellis.spark.io.file.cog.byteReader
 import geotrellis.util.ByteReader
-
-import com.typesafe.config.ConfigFactory
 import java.net.URI
+
+import geotrellis.spark.io.file.conf.FileConfig
 
 case class FileGeoTiffLayerReader[M[T] <: Traversable[T]](
   attributeStore: AttributeStore[M, GeoTiffMetadata],
@@ -22,5 +21,5 @@ case class FileGeoTiffLayerReader[M[T] <: Traversable[T]](
 }
 
 object FileGeoTiffLayerReader {
-  val defaultThreadCount: Int = ConfigFactory.load().getThreads("geotrellis.file.threads.collection.read")
+  val defaultThreadCount: Int = FileConfig.threads.collection.readThreads
 }