+++ spark/src/main/scala/geotrellis/spark/io/hadoop/conf/CamelCaseConfig.scala
@@ -1,0 +1,7 @@
+package geotrellis.spark.io.hadoop.conf
+
+import pureconfig.{CamelCase, ConfigFieldMapping, ProductHint}
+
+trait CamelCaseConfig {
+  implicit def hint[T]: ProductHint[T] = ProductHint[T](ConfigFieldMapping(CamelCase, CamelCase))
+}